Split out zoomed-in "NodeWithBody" from "NodeGraphView"
