+++ loom-editor/src/components/NodeGraphView/index.test.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { renderWithProvider } from "../../utils/test-utils";
-import { screen, fireEvent } from "@testing-library/react";
+import { screen } from "@testing-library/react";
 
 import NodeGraphView from "./";
 import { YarnGraphNode } from "../NodeGraph";
@@ -17,38 +17,6 @@ describe("<NodeGraphView />", () => {
 
   it("renders", () => {
     renderWithProvider(<NodeGraphView node={mockNode} />);
-  });
-
-  it("renders no tags if there are none", () => {
-    const nodeWithNoTags = {
-      ...mockNode,
-      yarnNode: {
-        ...mockNode.yarnNode,
-        tags: "",
-      },
-    };
-    renderWithProvider(<NodeGraphView node={nodeWithNoTags} />);
-    expect(screen.queryByTestId("node-graph-view-tags")).toBeNull();
-  });
-
-  it("opens the color chooser", () => {
-    renderWithProvider(<NodeGraphView node={mockNode} />);
-
-    expect(screen.queryByTestId("node-title-color-chooser")).toBeNull();
-
-    fireEvent.click(screen.getByTitle("Change node color"));
-
-    expect(screen.queryByTestId("node-title-color-chooser")).not.toBeNull();
-  });
-
-  it("opens the tag chooser", () => {
-    renderWithProvider(<NodeGraphView node={mockNode} />);
-
-    expect(screen.queryByTestId("node-tag-chooser")).toBeNull();
-
-    fireEvent.click(screen.getByTitle("Manage node tags"));
-
-    expect(screen.queryByTestId("node-tag-chooser")).not.toBeNull();
   });
 
   describe("searching", () => {
